+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/skpr/cognito-auth/cmd"
-	"github.com/skpr/cognito-auth/cmd/google"
+	"github.com/skpr/cognito-auth/cmd/oidc"
 	"github.com/skpr/cognito-auth/cmd/userpool"
 	"gopkg.in/alecthomas/kingpin.v2"
 	"os"
@@ -12,8 +12,8 @@ func main() {
 
 	app := kingpin.New("cognito-auth", "Cognito CLI authentication commands")
 
-	cmdGoogle := app.Command("google", "Google commands").Alias("g")
-	google.Login(cmdGoogle)
+	cmdOidc := app.Command("oidc", "OpenID Connect commands")
+	oidc.Login(cmdOidc)
 
 	cmdUserpool := app.Command("userpool", "Userpool commands").Alias("up")
 	userpool.Login(cmdUserpool)
