+++ pkg/oidc/tokens_refresher.go
@@ -1,4 +1,4 @@
-package googleauth
+package oidc
 
 import (
 	"context"
